requirement.txt
pycryptodome
pillow
supabase
python-dotenv
pysqleet